requirements.txt
annotated-types=0.7.0
anyio=6.2.post1
blinker=1.9.0
cachetools=5.5.0
certifi=2024.8.30
charset-normalizer=3.4.0
click=8.1.7
colorama=0.4.6
distro=1.9.0
Flask=3.1.0
google-api-core=2.23.0
google-api-python-client=2.153.0
google-auth=2.36.0
google-auth-httplib2=0.2.0
google-auth-oauthlib=1.2.1
googleapis-common-protos=1.66.0
groq=0.12.0
h11=0.14.0
httpcore=1.0.7
httplib2=0.22.0
httpx=0.27.2
idna=3.10
itsdangerous=2.2.0
Jinja2=3.1.4
jiter=0.7.1
MarkupSafe=3.0.2
numpy=2.1.3
oauthlib=3.2.2
openai=1.54.5
pandas=2.2.3
pip=24.3.1
proto-plus=1.25.0
protobuf=5.28.3
pyasn1=0.6.1
pyasn1_modules=0.4.1
pydantic=2.9.2
pydantic_core=2.23.4
pyparsing=3.2.0
python-dateutil=2.9.0.post0
python-dotenv=1.0.1
pytz=2024.2
requests=2.32.3
requests-oauthlib=2.0.0
rsa=4.9
six=1.16.0
sniffio=1.3.1
tqdm=4.67.0
typing_extensions=4.12.2
tzdata=2024.2
uritemplate=4.1.1
urllib3=2.2.3
Werkzeug=3.1.3